Write the rotation used on the caps too
This way we can solve the case of showing a prerolled frame
with a rotation. GstVideoSink always calls the show_frame virtual
function in the GstBaseSink preroll virtual function implementation
but if we only use the tag event to rotate the frame on the sink, the
events are sent after the prerolled frame has been shown and thus
there is no rotation on the sink, only after the events are recevied.
Setting the rotation on the caps make the prerolled buffer to be
shown correctly.