Keep custom events until the streams are available
Any non-handled event goes to the default pad's event handler.
The default approach is to push on every element's opposite pad
but it is possible that no pad exists yet and thus the event is lost.
Now, for custom events, if that happens, the events are first
stored and once the streams are available they are pushed downstream.