Fix wrong normalization of mp4 ctts offsets
The mp4 states that ctts offset is an unsigned value. But regardless of being
coded as unsigned, it is handled as signed (qtdemux 0.1x, qtdemux 1.5, adobe
media server, mozilla media player, etc). Analyzed some mp4 files, many standard
ones are muxed according to this interpretation, i.e., the key frames have a
ctts offset of 0, while others have positive or negative values (always encoded
as unsigned anyway) to produce correct presentation timestamps.

But there are some muxers that normalize these values so that when converting
the values to signed the result is always positive, thus interpreting the
specification differently. These cause a unintended time shift when computing
the presentation timestamps according to the specification. It is clear that it
is unintended because it actually breaks synch across different streams. There
is also a script to “fix” files produced by such muxers that reset ctts offsets
to 0 in order to fix the broken synch.

The h264 decoders tested do not actually need this offset applied as they
generate the correct timestamps. We have verified that simply ignoring the ctts
offset fixed the issue, the timestamps output by the decoder were correct and
without any undesired time shift. But we intend to fix this while still
producing correct presentation timestamps from the demuxer even if the decoder
does not need them. This will minimize any risk if eventually any downstream
element relies on receiving correct presentation timestamps from the demuxer.
The agreed solution is to store the ctts offset of key frames, and to
subtract it to all ctts offsets of all frames, including key frames. This
produces ctts offsets normalized to the most widely accepted interpretation
as explained above, and will only affect files muxed with “bad” ctts offsets
that would produce the undesired time shift.