Normalize ct offset also in qtdemux_parse_trun
Depending on the mp4 input, the ct offset are parsed in qtdemux_parse_samples
or qtdemux_parse_trun. A previous commit fixed qtdemux_parse_samples, and this
ones applies exactly the same to qtdemux_parse_trun.

This normalization is a workaround for some mp4 files that have a wrong offset
applied to ct offset due to a misinterpretation of the spec, in order to
null the unintended delay introduced by it that break sync. Even if it is a
problem of the mp4 file and not the demuxed, we needed to pass tests with such
bad files that we cannot change. Also this type of error in mp4 files is not
very uncommon and it is very convenient to be able to handle it in the demuxer.