Change default souphttpsource timeout to 10s
Was changed from infinite to 30s a long time ago. It was still too large and
needs to be shorter, like 10s. 10s is still a lot more than any alive server
should answer, specially for any media content.
We use a default play timeout of 30s, so if server was down, we finished with
play timeout error instead of the more specific error OPEN_FAILED from
souphttpsource.